#3485: ironing (remove getter from interface)
